python/preview/dyn_tool_generation_forecasting/requirements.txt
ag-ui-protocol==0.1.10
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
alembic==1.17.2
annotated-doc==0.0.4
annotated-types==0.7.0
anthropic==0.75.0
anyio==4.12.0
argcomplete==3.6.3
asyncer==0.0.8
attrs==25.4.0
Authlib==1.6.6
backoff==2.2.1
beartype==0.22.9
bleak==2.1.0
blinker==1.9.0
boto3==1.42.16
botocore==1.42.16
cachetools==6.2.4
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
cloudpickle==3.1.2
cohere==5.20.1
colorama==0.4.6
colorlog==6.10.1
contourpy==1.3.3
cryptography==46.0.3
cycler==0.12.1
cyclopts==4.4.1
diskcache==5.6.3
distro==1.9.0
dnspython==2.8.0
docstring_parser==0.17.0
docutils==0.22.4
dspy==3.0.4
email-validator==2.3.0
eval_type_backport==0.3.1
exceptiongroup==1.3.1
executing==2.2.1
fakeredis==2.33.0
fasta2a==0.6.0
fastapi==0.127.0
fastavro==1.12.1
fastmcp==2.14.1
fastuuid==0.14.0
filelock==3.20.1
Flask==3.1.2
fonttools==4.61.1
frozenlist==1.8.0
fsspec==2025.12.0
genai-prices==0.0.49
gepa==0.0.17
google-auth==2.45.0
google-genai==1.56.0
googleapis-common-protos==1.72.0
griffe==1.15.0
groq==1.0.0
grpcio==1.67.1
h11==0.16.0
hf-xet==1.2.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.3
huggingface-hub==0.36.0
idna==3.11
importlib_metadata==8.7.1
invoke==2.2.1
itsdangerous==2.2.0
jaraco.classes==3.4.0
jaraco.context==6.0.2
jaraco.functools==4.4.0
Jinja2==3.1.6
jiter==0.12.0
jmespath==1.0.1
joblib==1.5.3
json_repair==0.54.3
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.25.1
jsonschema-path==0.3.4
jsonschema-specifications==2025.9.1
keyring==25.7.0
kiwisolver==1.4.9
langchain==1.2.6
langchain-core==1.2.7
langchain-openai==1.1.6
langgraph==1.0.5
langgraph-checkpoint==3.0.1
langgraph-prebuilt==1.0.5
langgraph-sdk==0.3.1
langsmith==0.5.0
litellm==1.80.11
logfire==4.16.0
logfire-api==4.16.0
lupa==2.6
magicattr==0.1.6
Mako==1.3.10
markdown-it-py==4.0.0
MarkupSafe==3.0.3
matplotlib==3.10.8
mcp==1.25.0
mdurl==0.1.2
mistralai==1.9.11
more-itertools==10.8.0
mpmath==1.3.0
multidict==6.7.0
nest-asyncio==1.6.0
networkx==3.6.1
nexus-rpc==1.2.0
numpy==2.4.0
ollama==0.6.1
openai==2.14.0
openapi-pydantic==0.5.1
openevolve==0.2.24
opentelemetry-api==1.39.1
opentelemetry-exporter-otlp-proto-common==1.39.1
opentelemetry-exporter-otlp-proto-http==1.39.1
opentelemetry-exporter-prometheus==0.60b1
opentelemetry-instrumentation==0.60b1
opentelemetry-instrumentation-httpx==0.60b1
opentelemetry-proto==1.39.1
opentelemetry-sdk==1.39.1
opentelemetry-semantic-conventions==0.60b1
opentelemetry-util-http==0.60b1
optuna==4.6.0
orjson==3.11.5
ormsgpack==1.12.1
packaging==25.0
pandas==2.3.3
pandas-stubs==2.3.3.251219
pathable==0.4.4
pathvalidate==3.3.1
patsy==1.0.2
pillow==12.0.0
platformdirs==4.5.1
prometheus_client==0.23.1
prompt_toolkit==3.0.52
propcache==0.4.1
protobuf==6.33.2
py-key-value-aio==0.3.0
py-key-value-shared==0.3.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
pydantic==2.12.5
pydantic-ai==1.39.0
pydantic-ai-slim==1.39.0
pydantic-evals==1.39.0
pydantic-graph==1.39.0
pydantic-settings==2.12.0
pydantic_core==2.41.5
pydocket==0.16.3
Pygments==2.19.2
PyJWT==2.10.1
pyobjc-core==12.1
pyobjc-framework-Cocoa==12.1
pyobjc-framework-CoreBluetooth==12.1
pyobjc-framework-libdispatch==12.1
pyparsing==3.2.5
pyperclip==1.11.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-json-logger==4.0.0
python-multipart==0.0.21
pytz==2025.2
PyYAML==6.0.3
redis==7.1.0
referencing==0.36.2
regex==2025.11.3
requests==2.32.5
requests-toolbelt==1.0.0
rich==14.2.0
rich-rst==1.3.2
rpds-py==0.30.0
rsa==4.9.1
s3transfer==0.16.0
scikit-learn==1.8.0
scipy==1.16.3
scipy-stubs==1.16.3.3
setuptools==80.9.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
SQLAlchemy==2.0.45
sse-starlette==3.0.4
starlette==0.50.0
statsmodels==0.14.6
sympy==1.14.0
tabulate==0.9.0
temporalio==1.20.0
tenacity==9.1.2
threadpoolctl==3.6.0
tiktoken==0.12.0
tokenizers==0.22.1
torch==2.9.1
tqdm==4.67.1
typer==0.20.1
types-protobuf==6.32.1.20251210
types-requests==2.32.4.20250913
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.2
uuid_utils==0.12.0
uv==0.9.18
uvicorn==0.40.0
wcwidth==0.2.14
websockets==15.0.1
Werkzeug==3.1.4
wrapt==1.17.3
xxhash==3.6.0
yarl==1.22.0
zipp==3.23.0
zstandard==0.25.0
